Former-commit-id: 269d2923a97e3fd13651e54a0921f117136ef747

+++ ide/src/edu/wustl/cse/lookingglass/apis/walkandtouch/animation/WalkToAnimation.java
@@ -306,7 +306,7 @@ public class WalkToAnimation extends AbstractWalkAnimation {
 			numberOfSteps = java.lang.Math.round(distanceToMove/stepLength);
 		}
 	
-		return distanceToMove/((double)numberOfSteps);
+		return distanceToMove/numberOfSteps;
 	
 	}
 }
